requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
allure-behave==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
behave==1.2.6
parse-type==0.6.0
parse==1.19.0
pluggy==1.0.0
pyhamcrest==2.0.3
six==1.16.0
